requirements.txt
annotated-doc==0.0.4
alembic==1.18.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.0.1
click==8.3.2
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
idna==3.11
Mako==1.3.11
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.2
PyJWT==2.10.1
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
